Impor dataset

Fungsi untuk mengimpor dataset hasil pencarian dari fungsi cari().

impor(.data)

Arguments

.data
data keluaran dari fungsi cari() (data dengan kelas bdgjwr)

Source

Dataset bersumber dari Open Data Kota Bandung http://data.bandung.go.id.

Value

dataframe atau list berisi dataframe apabila dataset yang hendak diimpor jumlahnya lebih dari satu.

Examples

library(bandungjuara)

kebakaran <- cari(kata_kunci = "kebakaran")

#> Berhasil menemukan 23 dataset dengan dengan kata kunci: kebakaran

kebakaran

#> # A tibble: 23 x 5
#>    nama             deskripsi            url             dibuat     diperbaharui
#>    <chr>            <chr>                <chr>           <date>     <date>      
#>  1 Data Kebakaran … Data ini berisi kol… http://data.ba… 2019-01-10 2019-01-10  
#>  2 Data Kebakaran … Data mengenai benca… http://data.ba… 2017-07-25 2018-08-29  
#>  3 Data Serapan An… Data ini berisi kol… http://data.ba… 2018-04-20 2018-04-20  
#>  4 Data Serapan An… Data ini berisi kol… http://data.ba… 2018-04-20 2018-04-20  
#>  5 Data Serapan An… Data ini berisi kol… http://data.ba… 2018-04-20 2018-04-20  
#>  6 Data Serapan An… Data ini berisi kol… http://data.ba… 2018-04-20 2018-04-20  
#>  7 Data Serapan An… Data ini berisi kol… http://data.ba… 2018-04-13 2018-04-13  
#>  8 Data Kebakaran … Data ini berisi kol… http://data.ba… 2018-01-15 2018-01-15  
#>  9 Data Kebakaran … Data mengenai benca… http://data.ba… 2017-07-25 2018-01-04  
#> 10 Data Rekrutmen … Dataset tentang Dat… https://git.ba… 2017-07-25 NA          
#> # … with 13 more rows

dataset_kebakaran <- impor(kebakaran)

#> Data yang tidak berhasil diunduh:

#> Data Kebakaran Kota Bandung Tahun 2015
#> Data Bantuan Kebakaran Luar Kota Bandung Tahun 2015

dataset_kebakaran

#> $`Data Kebakaran Kota Bandung Tahun 2018`
#> # A tibble: 12 x 23
#>       no bulan jumlah_kebakaran pbk_bp pbk_psr pbk_bi pbk_sc pbk_bu
#>    <dbl> <chr>            <dbl>  <dbl> <chr>   <chr>  <chr>  <chr> 
#>  1     1 Janu…               19      8 -       1      -      1     
#>  2     2 Febr…               19      8 -       -      -      4     
#>  3     3 Maret                9      1 -       -      -      3     
#>  4     4 April               12      6 -       -      -      5     
#>  5     5 Mei                 19      8 -       1      -      2     
#>  6     6 Juni                24      9 -       -      -      4     
#>  7     7 Juli                42     12 1       1      -      7     
#>  8     8 Agus…               20      7 -       1      -      -     
#>  9     9 Sept…               34      8 1       2      -      5     
#> 10    10 Okto…               35      8 1       -      -      5     
#> 11    11 Nove…                9      4 -       -      -      1     
#> 12    12 Dese…               15      5 -       -      -      4     
#> # … with 15 more variables: pbk_lain_lain <dbl>, pk_kp <chr>, pk_lp <chr>,
#> #   pk_ls <chr>, pk_rk <chr>, pk_ln <chr>, jt_bp <dbl>, jt_kr <chr>,
#> #   jt_bi <chr>, jt_bu <chr>, jt_sc <chr>, jt_kios <chr>, jt_ln <chr>,
#> #   korban_md <chr>, korban_lk <chr>
#> 
#> $`Data Serapan Anggaran Dinas Pencegahan dan Penanggulangan Kebakaran Tahun 2013`
#> # A tibble: 12 x 8
#>    skpd    program  kegiatan   anggaran realisasi persentase_sera… tahun catatan
#>    <chr>   <chr>    <chr>         <dbl>     <dbl>            <dbl> <dbl> <chr>  
#>  1 Dinas … Program… Pengadaan…   2.50e9    2.44e9             97.7  2013 -      
#>  2 Dinas … Program… Penyuluha…   1.03e9    9.66e8             93.6  2013 -      
#>  3 Dinas … Program… Pemelihar…   9.04e8    7.85e8             86.8  2013 -      
#>  4 Dinas … Program… Peningkat…   7.87e8    7.79e8             99.0  2013 -      
#>  5 Dinas … Program… Rekrutmen…   7.83e8    7.38e8             94.3  2013 -      
#>  6 Dinas … Program… Rehabilit…   3.63e8    3.50e8             96.5  2013 -      
#>  7 Dinas … Program… Sosialisa…   9.19e7    8.71e7             94.8  2013 -      
#>  8 Dinas … Program… Pencegaha…   7.91e7    7.79e7             98.5  2013 -      
#>  9 Dinas … Program… Pengawasa…   7.48e7    5.79e7             77.4  2013 -      
#> 10 Dinas … Program… Monitorin…   5.51e7    5.19e7             94.2  2013 -      
#> 11 Dinas … Program… Penyusuna…   5.23e7    4.46e7             85.2  2013 -      
#> 12 Dinas … Program… Pendidika…   4.32e7    3.76e7             87.0  2013 -      
#> 
#> $`Data Serapan Anggaran Dinas Pencegahan dan Penanggulangan Kebakaran Tahun 2014`
#> # A tibble: 12 x 8
#>    skpd    program  kegiatan   anggaran realisasi persentase_sera… tahun catatan
#>    <chr>   <chr>    <chr>         <dbl>     <dbl>            <dbl> <dbl> <chr>  
#>  1 Dinas … Program… "Pengadaa…   8.25e9    8.20e9             99.4  2014 -      
#>  2 Dinas … Program… Pencegaha…   2.03e9    1.95e9             95.8  2014 -      
#>  3 Dinas … Program… Peningkat…   1.17e9    1.16e9             98.4  2014 -      
#>  4 Dinas … Program… Pemelihar…   1.07e9    1.01e9             94.8  2014 -      
#>  5 Dinas … Program… Rehabilit…   6.29e8    6.16e8             97.9  2014 -      
#>  6 Dinas … Program… Penyuluha…   3.71e8    3.60e8             97.2  2014 -      
#>  7 Dinas … Program… Rekrutmen…   3.25e8    3.01e8             92.5  2014 -      
#>  8 Dinas … Program… Sosialisa…   2.54e8    2.26e8             89.0  2014 -      
#>  9 Dinas … Program… Pengawasa…   2.44e8    2.17e8             88.7  2014 -      
#> 10 Dinas … Program… Penyusuna…   1.66e8    1.03e8             62.0  2014 -      
#> 11 Dinas … Program… Pendidika…   1.20e8    1.03e8             85.7  2014 -      
#> 12 Dinas … Program… Monitorin…   8.29e7    7.35e7             88.6  2014 -      
#> 
#> $`Data Serapan Anggaran Dinas Pencegahan dan Penanggulangan Kebakaran Tahun 2015`
#> # A tibble: 12 x 8
#>    skpd    program  kegiatan   anggaran realisasi persentase_sera… tahun catatan
#>    <chr>   <chr>    <chr>         <dbl>     <dbl>            <dbl> <dbl> <chr>  
#>  1 Dinas … Program… Pengadaan…   5.85e9    5.81e9             99.3  2015 -      
#>  2 Dinas … Program… Peningkat…   1.55e9    1.55e9             99.8  2015 -      
#>  3 Dinas … Program… Pencegaha…   1.49e9    1.49e9             99.6  2015 -      
#>  4 Dinas … Program… Pemelihar…   9.39e8    9.22e8             98.2  2015 -      
#>  5 Dinas … Program… Rehabilit…   5.10e8    5.03e8             98.5  2015 -      
#>  6 Dinas … Program… Pengawasa…   4.02e8    3.94e8             97.9  2015 -      
#>  7 Dinas … Program… Penyuluha…   3.37e8    3.33e8             98.8  2015 -      
#>  8 Dinas … Program… Sosialisa…   2.37e8    2.33e8             98.3  2015 -      
#>  9 Dinas … Program… Rekrutmen…   1.92e8    1.78e8             93.1  2015 -      
#> 10 Dinas … Program… Pendidika…   1.84e8    1.81e8             98.5  2015 -      
#> 11 Dinas … Program… Monitorin…   1.72e8    1.67e8             97.1  2015 -      
#> 12 Dinas … Program… Penyusuna…   3.78e7    3.78e7            100.0  2015 -      
#> 
#> $`Data Serapan Anggaran Dinas Pencegahan dan Penanggulangan Kebakaran Tahun 2016`
#> # A tibble: 12 x 8
#>    skpd    program  kegiatan   anggaran realisasi persentase_sera… tahun catatan
#>    <chr>   <chr>    <chr>         <dbl>     <dbl>            <dbl> <dbl> <chr>  
#>  1 Dinas … Program… Pengadaan…   8.16e9    7.34e9             89.9  2016 -      
#>  2 Dinas … Program… Kegiatan …   2.83e9    2.61e9             92.2  2016 -      
#>  3 Dinas … Program… Peningkat…   1.99e9    1.98e9             99.5  2016 -      
#>  4 Dinas … Program… Pemelihar…   1.69e9    1.53e9             90.2  2016 -      
#>  5 Dinas … Program… Kegiatan …   1.52e9    1.47e9             96.3  2016 -      
#>  6 Dinas … Program… Rehabilit…   1.05e9    1.01e9             96.4  2016 -      
#>  7 Dinas … Program… Kegiatan …   8.75e8    8.35e8             95.5  2016 -      
#>  8 Dinas … Program… Pengawasa…   6.90e8    6.54e8             94.8  2016 -      
#>  9 Dinas … Program… Kegiatan …   4.72e8    4.60e8             97.4  2016 -      
#> 10 Dinas … Program… Monitorin…   4.70e8    4.35e8             92.5  2016 -      
#> 11 Dinas … Program… Penyusuna…   1.91e8    1.84e8             96.3  2016 -      
#> 12 Dinas … Program… Sosialisa…   1.11e8    1.09e8             97.8  2016 -      
#> 
#> $`Data Serapan Anggaran Dinas Kebakaran dan Penanggulangan Bencana Tahun 2017`
#> # A tibble: 16 x 8
#>    skpd    program  kegiatan   anggaran realisasi persentase_sera… tahun catatan
#>    <chr>   <chr>    <chr>         <dbl>     <dbl>            <dbl> <dbl> <chr>  
#>  1 Dinas … Program… Kegiatan …   9.29e9    8.64e9             93.0  2017 -      
#>  2 Dinas … Program… Kegiatan …   2.56e9    2.49e9             97.0  2017 -      
#>  3 Dinas … Program… Kegiatan …   1.25e9    1.19e9             95.2  2017 -      
#>  4 Dinas … Program… Kegiatan …   7.45e8    7.24e8             97.2  2017 -      
#>  5 Dinas … Program… Kegiatan …   5.57e8    5.33e8             95.6  2017 -      
#>  6 Dinas … Program… Kegiatan …   4.07e8    4.00e8             98.3  2017 -      
#>  7 Dinas … Program… Kegiatan …   1.65e8    1.54e8             93.0  2017 -      
#>  8 Dinas … Program… Kegiatan …   1.59e8    1.56e8             98    2017 -      
#>  9 Dinas … Program… Kegiatan …   1.58e8    1.58e8            100.0  2017 -      
#> 10 Dinas … Program… Kegiatan …   1.57e8    1.56e8             99.9  2017 -      
#> 11 Dinas … Program… Kegiatan …   1.47e8    1.47e8            100.0  2017 -      
#> 12 Dinas … Program… Kegiatan …   1.27e8    1.27e8             99.6  2017 -      
#> 13 Dinas … Program… Kegiatan …   1.27e8    1.26e8             99.3  2017 -      
#> 14 Dinas … Program… Kegiatan …   1.09e8    1.09e8            100    2017 -      
#> 15 Dinas … Program… Kegiatan …   7.00e7    6.96e7             99.5  2017 -      
#> 16 Dinas … Program… Kegiatan …   6.00e7    5.78e7             96.2  2017 -      
#> 
#> $`Data Kebakaran Kota Bandung Tahun 2017`
#> # A tibble: 195 x 21
#>    hari  tanggal waktu_awal waktu_akhir lokasi_kejadian kelurahan kecamatan
#>    <chr> <chr>   <chr>      <chr>       <chr>           <chr>     <chr>    
#>  1 Ming… 01-Jan… 11.36      12.08       Jl.Sekelimus 4… Batunung… Bandung …
#>  2 Ming… 01-Jan… 0.40       3.00        Jl.Cigondewah … Cigondew… Bandung …
#>  3 Ming… 01-Jan… 16.30      17.03       Jl.Setra Ria N… Sukasari  Sukasari 
#>  4 Ming… 01-Jan… 21.00      21.02       Jl.Paledang Rt… Campaka   Andir    
#>  5 Sela… 03-Jan… 5.45       7.03        Jl.Moch. Toha … Karasak   Astana A…
#>  6 Sela… 03-Jan… 18.02      19.45       Jl.Tanjung Sar… Antapani… Antapani 
#>  7 Rabu  04-Jan… 21.15      21.02       Jl.Layang Paso… Taman sa… Bandung …
#>  8 Jumat 06-Jan… 6.04       7.03        Jl.Buah Batu (… Mengger   Bandung …
#>  9 Senin 09-Jan… 22.03      5.03        Jl. Jend.Sudir… Ciroyom   Andir    
#> 10 Sela… 10-Jan… 12.45      14.04       Jl. Jend.Sudir… Ciroyom   Andir    
#> # … with 185 more rows, and 14 more variables: benda_yang_terbakar <chr>,
#> #   bangunan_yang_terselamatkan <chr>, nama_pemilik <chr>,
#> #   penyebab_kebakaran <chr>, korban_meninggal_dunia <chr>, korban_luka <chr>,
#> #   mobil_pancar <chr>, mobil_rescue <chr>,
#> #   mobil_unit_lain_yang_bertugas_selain_pancar_dan_rescue <chr>, apar <chr>,
#> #   sumber_air_pdam_m3 <chr>, sumber_air_lainnya <chr>,
#> #   taksiran_kerugian <chr>, luas_areal_yang_terbakar_m2 <chr>
#> 
#> $`Data Kebakaran Kota Bandung Tahun 2016`
#> # A tibble: 106 x 16
#>    tanggal waktu_awal waktu_akhir lokasi_kebakaran kelurahan kecamatan
#>    <chr>   <chr>      <chr>       <chr>            <chr>     <chr>    
#>  1 02/01/… 14.20      15.35       JL. INDUSTRI RT… Arjuna    Cicendo  
#>  2 04/01/… 7.40       8.15        PERUM SARIJADI … Sukawarna Sukajadi 
#>  3 07/01/… 19.15      19.50       JL. KEBON JAYAN… Kebon Ja… Kiaracon…
#>  4 08/01/… 19.55      22.45       JL. DAGO         Ciharala… Cimenyan 
#>  5 12/01/… 1.40       1.55        JL. FLORES NO.08 Citarum   Bandung …
#>  6 18/01/… 13.35      4.30        PADASUKA NO.118… Pasir La… Cibeunyi…
#>  7 30/01/… 9.45       10.30       JL. CIBODAS RAY… Antapani… Antapani 
#>  8 30/01/… 3.25       5.15        JL. SEKEMIRUNG … Cigadung  Cibeunyi…
#>  9 30/01/… 15.30      16.00       JL. RY PASTEUR … Sukagalih Sukajadi 
#> 10 31/01/… 9.20       10.55       BOJONG RAYA HOL… Caringin  Bandung …
#> # … with 96 more rows, and 10 more variables: benda_yang_terbakar <chr>,
#> #   bangunan_yang_terselamatkan <chr>, penyebab_kebakaran <chr>,
#> #   korban_mati <chr>, korban_luka <chr>, sumber_air_pdam_m <chr>,
#> #   taksiran_kerugian <dbl>, luas_areal_yang_terbakar <chr>,
#> #   mobil_pancar <dbl>, mobil_rescue <chr>
#> 
#> $`Data Rekrutmen Relawan Satuan Relawan Kebakaran (Satwankar) Tahun 2014`
#> # A tibble: 19 x 3
#>    tanggal   kecamatan        jumlah
#>    <chr>     <chr>             <dbl>
#>  1 6/26/2014 Sukasari             17
#>  2 6/26/2014 Sukajadi             17
#>  3 6/26/2014 Andir                18
#>  4 6/26/2014 Cidadap              17
#>  5 6/26/2014 Coblong              17
#>  6 6/26/2014 Cibeunying Kaler     17
#>  7 6/24/2014 Bandung Wetan        32
#>  8 6/24/2014 Sumur Bandung        32
#>  9 6/24/2014 Batununggal          32
#> 10 6/24/2014 Cibeunying Kidul     32
#> 11 6/24/2014 Lengkong             32
#> 12 6/24/2014 Arcamanik            24
#> 13 6/24/2014 Rancasari            24
#> 14 6/24/2014 Cibiru               24
#> 15 6/24/2014 Panyileukan          24
#> 16 6/24/2014 Gede Bage            24
#> 17 6/24/2014 Ujung Berung         24
#> 18 6/24/2014 Mandalajati          24
#> 19 6/24/2014 Cinambo              24
#> 
#> $`Data Alat Pendukung Pemadam Kebakaran`
#> # A tibble: 17 x 3
#>    nama_alat                       jumlah satuan
#>    <chr>                            <dbl> <chr> 
#>  1 Alat Pemadam Api Besar               3 Buah  
#>  2 Breathing Aparatus                   9 Buah  
#>  3 Jaket Tahan Api                      4 set   
#>  4 Gun Nozle                           15 Buah  
#>  5 Selang Pancar                      300 Rol   
#>  6 Two  Way Conection                  12 Buah  
#>  7 Jaket Tahan Panas                   30 Set   
#>  8 Alat Pengisi Breathing Aparatus      2 Buah  
#>  9 Alat Pengukur Tekanan Air            2 Buah  
#> 10 Alat Poging                          2 Buah  
#> 11 Impeler Pump Floting Pump            4 Buah  
#> 12 Masker Full                        120 Buah  
#> 13 Selang Hisap                        60 Buah  
#> 14 Head Lamp                          120 Buah  
#> 15 Helmet Pemadam                     120 Buah  
#> 16 Gergaji Mesin                        8 Buah  
#> 17 Handy Talky                        100 Unit  
#> 
#> $`Rekapitulasi Data Kebakaran Luar Kota Bandung Pada Tahun 2007`
#> # A tibble: 12 x 27
#>    tahun bulan jumlah_kebakaran benda_yang_terb… benda_yang_terb…
#>    <dbl> <chr> <chr>            <chr>            <chr>           
#>  1  2007 JANU… -                -                -               
#>  2  2007 PEBR… 1                1                -               
#>  3  2007 MARET -                -                -               
#>  4  2007 APRIL 2                -                -               
#>  5  2007 MEI   -                -                -               
#>  6  2007 JUNI  1                -                -               
#>  7  2007 JULI  -                -                -               
#>  8  2007 AGUS… -                -                -               
#>  9  2007 SEPT… 3                1                -               
#> 10  2007 OKTO… 1                1                -               
#> 11  2007 NOPE… -                -                -               
#> 12  2007 DESE… -                -                -               
#> # … with 22 more variables: benda_yang_terbakar_bangunan_industri <chr>,
#> #   benda_yang_terbakar_shoping_centre <chr>,
#> #   benda_yang_terbakar_bangunan_umum <chr>,
#> #   benda_yang_terbakar_lain_lain <chr>, penyebab_kebakaran_kompor <chr>,
#> #   penyebab_kebakaran_lampu_tempel <chr>, penyebab_kebakaran_listrik <chr>,
#> #   penyebab_kebakaran_rokok <chr>, penyebab_kebakaran_lain_lain <chr>,
#> #   jumlah_yang_terbakar_bangunan_perumahan <chr>,
#> #   jumlah_yang_terbakar_kios <chr>,
#> #   jumlah_yang_terbakar_bangunan_industri <chr>,
#> #   jumlah_yang_terbakar_shoping_centre <chr>,
#> #   jumlah_yang_terbakar_lain_lain <chr>, korban_mati <chr>, korban_luka <chr>,
#> #   mobil_yang_bertugas <chr>, penggunaan_air_pdam_m3 <chr>,
#> #   penggunaan_air_lain_lain_m3 <chr>, areal_kebakaran_m2 <chr>,
#> #   bangunan_terselamatkan <chr>, taksiran_jumlah_kerugian_rp <chr>
#> 
#> $`Data Penyuluhan Pencegahan Kebakaran tahun 2015`
#> # A tibble: 12 x 13
#>    bulan pelajar_tk pelajar_sltp pelajar_slta pelajar_perguru… instansi_pemeri…
#>    <chr> <chr>      <chr>        <chr>        <chr>            <chr>           
#>  1 Janu… 289        NA           -            -                -               
#>  2 Febr… 2          -            -            -                -               
#>  3 Maret 2          -            75           -                -               
#>  4 April 843        100          100          -                370             
#>  5 Mei   436        -            -            500              75              
#>  6 Juni  358        1            -            -                150             
#>  7 Juli  -          -            470          -                -               
#>  8 Agus… 118        -            -            401              160             
#>  9 Sept… 196        300          -            100              150             
#> 10 Okto… 662        -            1            200              200             
#> 11 Nove… 1          -            -            500              340             
#> 12 Dese… 1          -            -            NA               50              
#> # … with 7 more variables: instansi_bumn <chr>, karyawan_hotel <chr>,
#> #   karyawan_rumah_sakit <chr>, pegawai_swasta <chr>, tni_polri <chr>,
#> #   kelurahan_kecamatan <chr>, ormas <chr>
#> 
#> $`Data Unit Kendaraan Pemadam Kebakaran`
#> # A tibble: 40 x 6
#>    tahun anggaran merk_mesin utility          merk_pompa  kondisi
#>    <dbl> <chr>    <chr>      <chr>            <chr>       <chr>  
#>  1  2001 APBD II  ISUZU      Mobil unit pompa NIHON KIKAI Baik   
#>  2  2001 APBD II  ISUZU      Mobil unit pompa NIHON KIKAI Baik   
#>  3  1983 APBD II  ISUZU      Mobil unit pompa NIHON KIKAI Baik   
#>  4  1997 APBD II  HINO       Mobil unit pompa WATEROUS    Baik   
#>  5  1997 APBD II  HINO       Mobil unit pompa WATEROUS    Baik   
#>  6  1997 DEPDAGRI MITSUBISHI Mobil unit pompa MORITA      Baik   
#>  7  1995 APBD II  HINO       Mobil unit pompa WATEROUS    Baik   
#>  8  1997 DEPDAGRI MITSUBISHI Mobil unit pompa MORITA      Baik   
#>  9  2000 APBD II  MITSUBISHI Mobil unit pompa MORITA      Baik   
#> 10  2003 BANGUB   ISUZU      Mobil unit pompa TOHATSU     Baik   
#> # … with 30 more rows
#> 
#> $`Rekapitulasi Data Kebakaran Luar Kota Bandung Pada Tahun 2008`
#> # A tibble: 12 x 27
#>    tahun bulan jumlah_kebakaran benda_yang_terb… benda_yang_terb…
#>    <dbl> <chr> <chr>            <chr>            <chr>           
#>  1  2008 JANU… -                -                -               
#>  2  2008 PEBR… 4                -                -               
#>  3  2008 MARET -                -                -               
#>  4  2008 APRIL 1                -                -               
#>  5  2008 MEI   1                1                -               
#>  6  2008 JUNI  -                -                -               
#>  7  2008 JULI  1                1                -               
#>  8  2008 AGUS… 4                3                -               
#>  9  2008 SEPT… 4                1                -               
#> 10  2008 OKTO… 4                1                -               
#> 11  2008 NOPE… -                -                -               
#> 12  2008 DESE… 1                1                -               
#> # … with 22 more variables: benda_yang_terbakar_bangunan_industri <chr>,
#> #   benda_yang_terbakar_shoping_centre <chr>,
#> #   benda_yang_terbakar_bangunan_umum <chr>,
#> #   benda_yang_terbakar_lain_lain <chr>, penyebab_kebakaran_kompor <chr>,
#> #   penyebab_kebakaran_lampu_tempel <chr>, penyebab_kebakaran_listrik <chr>,
#> #   penyebab_kebakaran_rokok <chr>, penyebab_kebakaran_lain_lain <chr>,
#> #   jumlah_yang_terbakar_bangunan_perumahan <chr>,
#> #   jumlah_yang_terbakar_kios <chr>,
#> #   jumlah_yang_terbakar_bangunan_industri <chr>,
#> #   jumlah_yang_terbakar_shoping_centre <chr>,
#> #   jumlah_yang_terbakar_lain_lain <chr>, korban_mati <chr>, korban_luka <chr>,
#> #   mobil_yang_bertugas <chr>, penggunaan_air_pdam_m3 <chr>,
#> #   penggunaan_air_lain_lain_m3 <chr>, areal_kebakaran_m2 <chr>,
#> #   bangunan_terselamatkan <chr>, taksiran_jumlah_kerugian_rp <chr>
#> 
#> $`Rekapitulasi Data Kebakaran Luar Kota Bandung Pada Tahun 2009`
#> # A tibble: 12 x 27
#>    tahun bulan jumlah_kebakaran benda_yang_terb… benda_yang_terb…
#>    <dbl> <chr> <chr>            <chr>            <chr>           
#>  1  2009 JANU… -                -                -               
#>  2  2009 FEBR… -                -                -               
#>  3  2009 MARET 2                2                -               
#>  4  2009 APRIL 1                1                -               
#>  5  2009 MEI   -                -                -               
#>  6  2009 JUNI  4                1                1               
#>  7  2009 JULI  1                1                -               
#>  8  2009 AGUS… 4                1                -               
#>  9  2009 SEPT… 3                1                -               
#> 10  2009 OKTO… 1                -                -               
#> 11  2009 NOPE… -                -                -               
#> 12  2009 DESE… 3                1                -               
#> # … with 22 more variables: benda_yang_terbakar_bangunan_industri <chr>,
#> #   benda_yang_terbakar_shoping_centre <chr>,
#> #   benda_yang_terbakar_bangunan_umum <chr>,
#> #   benda_yang_terbakar_lain_lain <chr>, penyebab_kebakaran_kompor <chr>,
#> #   penyebab_kebakaran_lampu_tempel <chr>, penyebab_kebakaran_listrik <chr>,
#> #   penyebab_kebakaran_rokok <chr>, penyebab_kebakaran_lain_lain <chr>,
#> #   jumlah_yang_terbakar_bangunan_perumahan <chr>,
#> #   jumlah_yang_terbakar_kios <chr>,
#> #   jumlah_yang_terbakar_bangunan_industri <chr>,
#> #   jumlah_yang_terbakar_shoping_centre <chr>,
#> #   jumlah_yang_terbakar_lain_lain <chr>, korban_mati <chr>, korban_luka <chr>,
#> #   mobil_yang_bertugas <chr>, penggunaan_air_pdam_m3 <chr>,
#> #   penggunaan_air_lain_lain_m3 <chr>, areal_kebakaran_m2 <chr>,
#> #   bangunan_terselamatkan <chr>, taksiran_jumlah_kerugian_rp <chr>
#> 
#> $`Rekapitulasi Data Kebakaran Luar Kota Bandung Pada Tahun 2010`
#> # A tibble: 12 x 27
#>    tahun bulan jumlah_kebakaran benda_yang_terb… benda_yang_terb…
#>    <dbl> <chr> <chr>            <chr>            <chr>           
#>  1  2010 JANU… -                -                -               
#>  2  2010 FEBR… 1                -                -               
#>  3  2010 MARET 1                -                -               
#>  4  2010 APRIL 2                1                -               
#>  5  2010 MEI   -                -                -               
#>  6  2010 JUNI  1                1                -               
#>  7  2010 JULI  2                1                1               
#>  8  2010 AGUS… -                -                -               
#>  9  2010 SEPT… -                -                -               
#> 10  2010 OKTO… 2                -                -               
#> 11  2010 NOPE… 1                1                -               
#> 12  2010 DESE… -                -                -               
#> # … with 22 more variables: benda_yang_terbakar_bangunan_industri <chr>,
#> #   benda_yang_terbakar_shoping_centre <chr>,
#> #   benda_yang_terbakar_bangunan_umum <chr>,
#> #   benda_yang_terbakar_lain_lain <chr>, penyebab_kebakaran_kompor <chr>,
#> #   penyebab_kebakaran_lampu_tempel <chr>, penyebab_kebakaran_listrik <chr>,
#> #   penyebab_kebakaran_rokok <chr>, penyebab_kebakaran_lain_lain <chr>,
#> #   jumlah_yang_terbakar_bangunan_perumahan <chr>,
#> #   jumlah_yang_terbakar_kios <chr>,
#> #   jumlah_yang_terbakar_bangunan_industri <chr>,
#> #   jumlah_yang_terbakar_shoping_centre <chr>,
#> #   jumlah_yang_terbakar_lain_lain <chr>, korban_mati <chr>, korban_luka <chr>,
#> #   mobil_yang_bertugas <chr>, penggunaan_air_pdam_m3 <chr>,
#> #   penggunaan_air_lain_lain_m3 <chr>, areal_kebakaran_m2 <chr>,
#> #   bangunan_terselamatkan <chr>, taksiran_jumlah_kerugian_rp <chr>
#> 
#> $`Rekapitulasi Data Kebakaran Luar Kota Bandung Pada Tahun 2011`
#> # A tibble: 12 x 27
#>    tahun bulan jumlah_kebakaran benda_yang_terb… benda_yang_terb…
#>    <dbl> <chr> <chr>            <chr>            <chr>           
#>  1  2011 JANU… 3                1                -               
#>  2  2011 FEBR… 1                -                -               
#>  3  2011 MARET 5                3                -               
#>  4  2011 APRIL -                -                -               
#>  5  2011 MEI   -                -                -               
#>  6  2011 JUNI  2                1                -               
#>  7  2011 JULI  1                1                -               
#>  8  2011 AGUS… 4                2                -               
#>  9  2011 SEPT… 4                1                -               
#> 10  2011 OKTO… -                -                -               
#> 11  2011 NOPE… -                -                -               
#> 12  2011 DESE… -                -                -               
#> # … with 22 more variables: benda_yang_terbakar_bangunan_industri <chr>,
#> #   benda_yang_terbakar_shoping_centre <chr>,
#> #   benda_yang_terbakar_bangunan_umum <chr>,
#> #   benda_yang_terbakar_lain_lain <chr>, penyebab_kebakaran_kompor <chr>,
#> #   penyebab_kebakaran_lampu_tempel <chr>, penyebab_kebakaran_listrik <chr>,
#> #   penyebab_kebakaran_rokok <chr>, penyebab_kebakaran_lain_lain <chr>,
#> #   jumlah_yang_terbakar_bangunan_perumahan <chr>,
#> #   jumlah_yang_terbakar_kios <chr>,
#> #   jumlah_yang_terbakar_bangunan_industri <chr>,
#> #   jumlah_yang_terbakar_shoping_centre <chr>,
#> #   jumlah_yang_terbakar_lain_lain <chr>, korban_mati <chr>, korban_luka <chr>,
#> #   mobil_yang_bertugas <chr>, penggunaan_air_pdam_m3 <chr>,
#> #   penggunaan_air_lain_lain_m3 <chr>, areal_kebakaran_m2 <chr>,
#> #   bangunan_terselamatkan <chr>, taksiran_jumlah_kerugian_rp <chr>
#> 
#> $`Rekapitulasi Data Kebakaran Luar Kota Bandung Pada Tahun 2012`
#> # A tibble: 12 x 27
#>    tahun bulan jumlah_kebakaran benda_yang_terb… benda_yang_terb…
#>    <dbl> <chr> <chr>            <chr>            <chr>           
#>  1  2012 JANU… -                -                -               
#>  2  2012 FEBR… -                -                -               
#>  3  2012 MARET -                -                -               
#>  4  2012 APRIL -                -                -               
#>  5  2012 MEI   3                -                -               
#>  6  2012 JUNI  -                -                -               
#>  7  2012 JULI  1                1                -               
#>  8  2012 AGUS… 2                -                -               
#>  9  2012 SEPT… 1                1                -               
#> 10  2012 OKTO… 2                -                -               
#> 11  2012 NOPE… 1                -                -               
#> 12  2012 DESE… -                -                -               
#> # … with 22 more variables: benda_yang_terbakar_bangunan_industri <chr>,
#> #   benda_yang_terbakar_shoping_centre <chr>,
#> #   benda_yang_terbakar_bangunan_umum <chr>,
#> #   benda_yang_terbakar_lain_lain <chr>, penyebab_kebakaran_kompor <chr>,
#> #   penyebab_kebakaran_lampu_tempel <chr>, penyebab_kebakaran_listrik <chr>,
#> #   penyebab_kebakaran_rokok <chr>, penyebab_kebakaran_lain_lain <chr>,
#> #   jumlah_yang_terbakar_bangunan_perumahan <chr>,
#> #   jumlah_yang_terbakar_kios <chr>,
#> #   jumlah_yang_terbakar_bangunan_industri <chr>,
#> #   jumlah_yang_terbakar_shoping_centre <chr>,
#> #   jumlah_yang_terbakar_lain_lain <chr>, korban_mati <chr>, korban_luka <chr>,
#> #   mobil_yang_bertugas <chr>, penggunaan_air_pdam_m3 <chr>,
#> #   penggunaan_air_lain_lain_m3 <chr>, areal_kebakaran_m2 <chr>,
#> #   bangunan_terselamatkan <chr>, taksiran_jumlah_kerugian_rp <chr>
#> 
#> $`Rekapitulasi Data Kebakaran Luar Kota Bandung Pada Tahun 2013`
#> # A tibble: 12 x 27
#>    tahun bulan jumlah_kebakaran benda_yang_terb… benda_yang_terb…
#>    <dbl> <chr> <chr>            <chr>            <chr>           
#>  1  2013 JANU… 1                -                -               
#>  2  2013 FEBR… -                -                -               
#>  3  2013 MARET -                -                -               
#>  4  2013 APRIL -                -                -               
#>  5  2013 MEI   1                1                -               
#>  6  2013 JUNI  3                1                -               
#>  7  2013 JULI  2                -                -               
#>  8  2013 AGUS… -                -                -               
#>  9  2013 SEPT… -                -                -               
#> 10  2013 OKTO… 1                -                -               
#> 11  2013 NOPE… -                -                -               
#> 12  2013 DESE… -                -                -               
#> # … with 22 more variables: benda_yang_terbakar_bangunan_industri <chr>,
#> #   benda_yang_terbakar_shoping_centre <chr>,
#> #   benda_yang_terbakar_bangunan_umum <chr>,
#> #   benda_yang_terbakar_lain_lain <chr>, penyebab_kebakaran_kompor <chr>,
#> #   penyebab_kebakaran_lampu_tempel <chr>, penyebab_kebakaran_listrik <chr>,
#> #   penyebab_kebakaran_rokok <chr>, penyebab_kebakaran_lain_lain <chr>,
#> #   jumlah_yang_terbakar_bangunan_perumahan <chr>,
#> #   jumlah_yang_terbakar_kios <chr>,
#> #   jumlah_yang_terbakar_bangunan_industri <chr>,
#> #   jumlah_yang_terbakar_shoping_centre <chr>,
#> #   jumlah_yang_terbakar_lain_lain <chr>, korban_mati <chr>, korban_luka <chr>,
#> #   mobil_yang_bertugas <chr>, penggunaan_air_pdam_m3 <chr>,
#> #   penggunaan_air_lain_lain_m3 <chr>, areal_kebakaran_m2 <chr>,
#> #   bangunan_terselamatkan <chr>, taksiran_jumlah_kerugian_rp <chr>
#> 
#> $`Rekapitulasi Data Kebakaran Luar Kota Bandung Pada Tahun 2014`
#> # A tibble: 12 x 27
#>    tahun bulan jumlah_kebakaran benda_yang_terb… benda_yang_terb…
#>    <dbl> <chr> <chr>            <chr>            <chr>           
#>  1  2014 JANU… -                -                -               
#>  2  2014 FEBR… -                -                -               
#>  3  2014 MARET -                -                -               
#>  4  2014 APRIL 1                1                -               
#>  5  2014 MEI   -                -                -               
#>  6  2014 JUNI  1                1                -               
#>  7  2014 JULI  -                -                -               
#>  8  2014 AGUS… -                -                -               
#>  9  2014 SEPT… 8                3                1               
#> 10  2014 OKTO… 8                3                -               
#> 11  2014 NOPE… 1                1                -               
#> 12  2014 DESE… -                -                -               
#> # … with 22 more variables: benda_yang_terbakar_bangunan_industri <chr>,
#> #   benda_yang_terbakar_shoping_centre <chr>,
#> #   benda_yang_terbakar_bangunan_umum <chr>,
#> #   benda_yang_terbakar_lain_lain <chr>, penyebab_kebakaran_kompor <chr>,
#> #   penyebab_kebakaran_lampu_tempel <chr>, penyebab_kebakaran_listrik <chr>,
#> #   penyebab_kebakaran_rokok <chr>, penyebab_kebakaran_lain_lain <chr>,
#> #   jumlah_yang_terbakar_bangunan_perumahan <chr>,
#> #   jumlah_yang_terbakar_kios <chr>,
#> #   jumlah_yang_terbakar_bangunan_industri <chr>,
#> #   jumlah_yang_terbakar_shoping_centre <chr>,
#> #   jumlah_yang_terbakar_lain_lain <chr>, korban_mati <chr>, korban_luka <chr>,
#> #   mobil_yang_bertugas <chr>, penggunaan_air_pdam_m3 <chr>,
#> #   penggunaan_air_lain_lain_m3 <chr>, areal_kebakaran_m2 <chr>,
#> #   bangunan_terselamatkan <chr>, taksiran_jumlah_kerugian_rp <chr>
#> 
#> $`Rekapitulasi Data Kebakaran Luar Kota Bandung Pada Tahun 2015`
#> # A tibble: 12 x 27
#>    tahun bulan jumlah_kebakaran benda_yang_terb… benda_yang_terb…
#>    <dbl> <chr> <chr>            <chr>            <chr>           
#>  1  2015 JANU… -                -                -               
#>  2  2015 FEBR… 2                -                -               
#>  3  2015 MARET 2                2                -               
#>  4  2015 APRIL 1                -                1               
#>  5  2015 MEI   -                -                -               
#>  6  2015 JUNI  -                -                -               
#>  7  2015 JULI  4                -                -               
#>  8  2015 AGUS… 4                -                -               
#>  9  2015 SEPT… 2                -                -               
#> 10  2015 OKTO… -                -                -               
#> 11  2015 NOPE… -                -                -               
#> 12  2015 DESE… -                -                -               
#> # … with 22 more variables: benda_yang_terbakar_bangunan_industri <chr>,
#> #   benda_yang_terbakar_shoping_centre <chr>,
#> #   benda_yang_terbakar_bangunan_umum <chr>,
#> #   benda_yang_terbakar_lain_lain <chr>, penyebab_kebakaran_kompor <chr>,
#> #   penyebab_kebakaran_lampu_tempel <chr>, penyebab_kebakaran_listrik <chr>,
#> #   penyebab_kebakaran_rokok <chr>, penyebab_kebakaran_lain_lain <chr>,
#> #   jumlah_yang_terbakar_bangunan_perumahan <chr>,
#> #   jumlah_yang_terbakar_kios <chr>,
#> #   jumlah_yang_terbakar_bangunan_industri <chr>,
#> #   jumlah_yang_terbakar_shoping_centre <chr>,
#> #   jumlah_yang_terbakar_lain_lain <chr>, korban_mati <chr>, korban_luka <chr>,
#> #   mobil_yang_bertugas <chr>, penggunaan_air_pdam_m3 <chr>,
#> #   penggunaan_air_lain_lain_m3 <chr>, areal_kebakaran_m2 <chr>,
#> #   bangunan_terselamatkan <chr>, taksiran_jumlah_kerugian_rp <chr>
#> 

impor(kebakaran[1,])

#> Semua dataset berhasil diunduh!

#> # A tibble: 12 x 23
#>       no bulan jumlah_kebakaran pbk_bp pbk_psr pbk_bi pbk_sc pbk_bu
#>    <dbl> <chr>            <dbl>  <dbl> <chr>   <chr>  <chr>  <chr> 
#>  1     1 Janu…               19      8 -       1      -      1     
#>  2     2 Febr…               19      8 -       -      -      4     
#>  3     3 Maret                9      1 -       -      -      3     
#>  4     4 April               12      6 -       -      -      5     
#>  5     5 Mei                 19      8 -       1      -      2     
#>  6     6 Juni                24      9 -       -      -      4     
#>  7     7 Juli                42     12 1       1      -      7     
#>  8     8 Agus…               20      7 -       1      -      -     
#>  9     9 Sept…               34      8 1       2      -      5     
#> 10    10 Okto…               35      8 1       -      -      5     
#> 11    11 Nove…                9      4 -       -      -      1     
#> 12    12 Dese…               15      5 -       -      -      4     
#> # … with 15 more variables: pbk_lain_lain <dbl>, pk_kp <chr>, pk_lp <chr>,
#> #   pk_ls <chr>, pk_rk <chr>, pk_ln <chr>, jt_bp <dbl>, jt_kr <chr>,
#> #   jt_bi <chr>, jt_bu <chr>, jt_sc <chr>, jt_kios <chr>, jt_ln <chr>,
#> #   korban_md <chr>, korban_lk <chr>